Part-1: Divided cart into Head, Body(Incomplete) and Footer. Tickets: BENTO-1079, BENTO-1080, BENTO-1081
