+++ src/pages/fileCentricCart/cartView.style.js
@@ -1,0 +1,158 @@
+export default () => ({
+  tableWrapper: {
+    margin: 'auto 3% auto 3%',
+    maxWidth: '100%',
+  },
+  customFooterStyle: {
+    background: '#f3f3f4',
+  },
+  tableTitleWizard: {
+    width: '400px',
+    float: 'right',
+    paddingTop: '8px',
+  },
+  myFilesWrapper: {
+    border: '#03A383 4px solid',
+    borderRadius: '35px',
+    marginTop: '200px',
+    marginBottom: '80px',
+    marginLeft: '3%',
+    marginRight: '3%',
+    paddingBottom: '20px',
+    background: 'white',
+  },
+  linkIcon: {
+    color: '#dc762f',
+    width: '20px',
+    verticalAlign: 'sub',
+    margin: '0px 0px 0px 2px',
+  },
+  helpIcon: {
+    verticalAlign: 'top',
+    width: '20px',
+    zIndex: '600',
+  },
+  topButtonGroup: {
+    textAlign: 'right',
+    padding: '10px 43px 15px 0px',
+    position: 'relative',
+  },
+  messageTop: {
+    position: 'absolute',
+    right: '33px',
+    top: '-128px',
+    zIndex: '400',
+  },
+  manifestButtonGroup: {
+    marginTop: '10px',
+    float: 'right',
+  },
+  downloadButton: {
+    height: '45px',
+    minWidth: '191px',
+    color: '#fff',
+    boxShadow: 'none',
+    backgroundColor: '#03A383',
+    padding: '6px 16px',
+    fontSize: '0.875rem',
+    boxSizing: 'border-box',
+    transition: 'background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,border 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms',
+    lineHeight: '1.75',
+    fontWeight: '500',
+    fontFamily: '"Roboto", "Helvetica", "Arial", sans-serif',
+    borderRadius: '10px',
+    textTransform: 'uppercase',
+    border: 'none',
+    verticalAlign: 'top',
+    marginTop: '6px',
+  },
+  popUpWindowText: {
+    fontFamily: 'Lato',
+    size: '16px',
+  },
+  okButton: {
+    background: '#98A19E',
+    color: '#fff',
+    cursor: 'pointer',
+    '&:hover': {
+      backgroundColor: 'rgba(152,161,158,0.6)',
+    },
+  },
+  cancelButton: {
+    background: '#42779A',
+    color: '#fff',
+    cursor: 'pointer',
+    '&:hover': {
+      backgroundColor: 'rgba(66,119,154,0.6)',
+    },
+  },
+  tableDeleteButton: {
+    background: '#fff',
+    border: '1px solid #ccc',
+    width: '29px',
+    cursor: 'pointer',
+    height: '26px',
+    borderRadius: '15%',
+    padding: '0',
+  },
+  tableDeleteButtonDiv: {
+  },
+  removeCell: {
+    cursor: 'pointer',
+    display: 'inline-flex',
+    outline: 'none',
+  },
+  removeThCell: {
+    top: '0px',
+    color: '#A61401',
+    zIndex: '100',
+    position: 'relative',
+    fontSize: '11pt',
+    borderTop: '#42779A 3px solid',
+    fontStyle: 'normal',
+    fontFamily: "'Lato Regular','Raleway', sans-serif",
+    fontWeight: 'bold',
+    paddingLeft: '20px',
+    borderBottom: '#42779A 3px solid',
+    letterSpacing: '0.06em',
+    backgroundColor: '#ffffff',
+    width: '120px',
+    textAlign: 'center',
+  },
+  removeHeadCell: {
+    cursor: 'pointer',
+    display: 'flex',
+    verticalAlign: 'top',
+  },
+  removeHeadCellText: {
+    display: 'inline-block',
+    cursor: 'default',
+    lineHeight: '37px',
+  },
+  removeHeadCellIcon: {
+    ursor: 'pointer',
+    display: 'flex',
+    marginTop: '1px',
+    verticalAlign: 'top',
+  },
+  removeHeadCellIconButton: {
+    color: '#A61401',
+    width: '25px',
+    marginTop: '5px',
+    height: '25px',
+  },
+  removeAllMessage: {
+    fontWeight: '500',
+    position: 'absolute',
+    top: '36px',
+    right: '0',
+    zIndex: '400',
+    background: '#fff',
+    border: '2px solid #A61401',
+    borderRadius: '7px',
+    fontSize: '12px',
+    width: '110px',
+    height: '48px',
+    padding: '5px 0px',
+  },
+});
