+++ src/pages/fileCentricCart/components/cartBody.style.js
@@ -1,0 +1,5 @@
+export default () => ({
+  tableStyle: {
+    maxWidth: '100%',
+  },
+});
