+++ src/pages/fileCentricCart/components/cartFooter.style.js
@@ -1,0 +1,18 @@
+export default () => ({
+  textField: {
+    minWidth: '412px',
+    marginRight: '10px',
+    resize: 'none',
+    borderRadius: '10px',
+    border: '1.5px solid #707070',
+    background: '#EBEBEB',
+    color: '#000',
+    fontFamily: 'Open Sans',
+    height: '170px',
+    padding: '15px',
+    fontSize: '10px',
+  },
+  manifestTextarea: {
+    marginTop: '20px',
+  },
+});
