+++ src/pages/fileCentricCart/components/cartHeader.style.js
@@ -1,0 +1,41 @@
+export default (theme) => ({
+  logo: {
+    position: 'absolute',
+    float: 'left',
+    marginTop: '-17.9px',
+    marginLeft: '-10px',
+    width: '101px',
+    filter: 'drop-shadow(-3px 2px 6px rgba(27,28,28,0.29))',
+  },
+  header: {
+    borderBottom: '#42779A 10px solid',
+    height: '77px',
+    maxWidth: '100%',
+    marginLeft: '3%',
+    marginRight: '3.05%',
+  },
+  headerTitle: {
+    maxWidth: theme.custom.maxContentWidth,
+    margin: 'auto',
+    float: 'left',
+    marginLeft: '85px',
+    paddingLeft: '3px',
+    marginBottom: '-30px',
+    position: 'absolute',
+  },
+  headerMainTitle: {
+    fontFamily: theme.custom.fontFamilySans,
+    fontWeight: '300',
+    letterSpacing: '0.017em',
+    color: '#03A383',
+    fontSize: '18pt',
+    lineHeight: '75px',
+    '& $headerMainTitleTwo': {
+      fontWeight: 'bold',
+      letterSpacing: '0.025em',
+    },
+  },
+  headerMainTitleTwo: {
+
+  },
+});
